burn investment coin on claim redemption
