+++ testdata/investment_marker.json
@@ -1,0 +1,30 @@
+{
+    "address": "tp18vmzryrvwaeykmdtu6cfrz5sau3dhc5c73ms0u",
+    "coins": [
+      {
+        "denom": "investment_coin",
+        "amount": "420"
+      }
+    ],
+    "public_key": "",
+    "account_number": 10,
+    "sequence": 0,
+    "permissions": [
+      {
+        "permissions": [
+          "burn",
+          "delete",
+          "deposit",
+          "admin",
+          "mint",
+          "withdraw"
+        ],
+        "address": "tp18vd8fpwxzck93qlwghaj6arh4p7c5n89x8kskz"
+      }
+    ],
+    "status": "active",
+    "denom": "investment_coin",
+    "total_supply": "420",
+    "marker_type": "coin",
+    "supply_fixed": false
+  }
